submitted_model/requirements.txt
wget
sox
libsndfile1
ffmpeg
text-unidecode
matplotlib>=3.3.2
git+https://github.com/NVIDIA/NeMo.git@main
nemo_toolkit[asr]
pandas
